pyproject.toml
[project]
name = "agentic-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "markdown>=3.10.2",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.0",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
    "weasyprint>=68.1",
]
